chassis: De-duplicate LogError and related methods
